Updated object manager tests.

+++ src/ray/test/run_object_manager_tests.sh
@@ -5,12 +5,42 @@
 # Cause the script to exit if a single command fails.
 set -e
 
-# Start Redis.
-./src/common/thirdparty/redis/src/redis-server --loglevel warning --loadmodule ./src/common/redis_module/libray_redis_module.so --port 6379 &
-sleep 1s
+# Get the directory in which this script is executing.
+SCRIPT_DIR="`dirname \"$0\"`"
+RAY_ROOT="$SCRIPT_DIR/../../.."
+RAY_ROOT="`( cd \"$RAY_ROOT\" && pwd )`"
+if [ -z "$RAY_ROOT" ] ; then
+  exit 1
+fi
+# Ensure we're in the right directory.
+if [ ! -d "$RAY_ROOT/python" ]; then
+  echo "Unable to find root Ray directory. Has this script moved?"
+  exit 1
+fi
 
-# || true allows failures, so cleanup properly executes
+CORE_DIR="$RAY_ROOT/python/ray/core"
+REDIS_DIR="$CORE_DIR/src/common/thirdparty/redis/src"
+REDIS_MODULE="$CORE_DIR/src/common/redis_module/libray_redis_module.so"
+STORE_EXEC="$CORE_DIR/src/plasma/plasma_store"
+
+echo "$STORE_EXEC"
+echo "$REDIS_DIR/redis-server --loglevel warning --loadmodule $REDIS_MODULE --port 6379"
+echo "$REDIS_DIR/redis-cli -p 6379 shutdown"
+
+# Allow cleanup commands to fail.
 killall plasma_store || true
-./src/ray/object_manager/object_manager_integration_test || true
+$REDIS_DIR/redis-cli -p 6379 shutdown || true
 
-./src/common/thirdparty/redis/src/redis-cli -p 6379 shutdown
+# Run tests.
+$REDIS_DIR/redis-server --loglevel warning --loadmodule $REDIS_MODULE --port 6379 &
+sleep 1s
+$CORE_DIR/src/ray/object_manager/object_manager_stress_test $STORE_EXEC
+$REDIS_DIR/redis-cli -p 6379 shutdown
+
+$REDIS_DIR/redis-server --loglevel warning --loadmodule $REDIS_MODULE --port 6379 &
+sleep 1s
+$CORE_DIR/src/ray/object_manager/object_manager_test $STORE_EXEC
+$REDIS_DIR/redis-cli -p 6379 shutdown
+
+# Include raylet integration test once it's ready.
+# $CORE_DIR/src/ray/raylet/object_manager_integration_test $STORE_EXEC
